Remove `store.py` in advisor
